test(db): fold the single-use suite into one test binary
CI's Test job died linking this crate's tests:

  collect2: fatal error: ld terminated with signal 7 [Bus error], core dumped

with RabbitMQ reporting 48 MB free on the same runner at the same second.
That is disk exhaustion, not a defect: ld mmaps its output, the write cannot
allocate, and touching the mapped page faults as SIGBUS rather than surfacing
a clean ENOSPC. There were zero test failures, zero panics and zero compile
errors in the log.

Not my code's fault, but my code's weight. Each integration-test file is a
separate linked binary, axiam-db already had 39 of them, and this PR added
two more. The runner's "Free up disk space" step already prunes dotnet,
Android and the tool cache, so the headroom that remained was what the new
binaries consumed.

Both new files are mine and both are about the same property — that a
single-use consume serialises — so they share a binary now instead of taking
one each. 39 back down to 38, all 14 tests kept, no coverage lost.

This trims the marginal cost rather than solving the underlying one: a crate
whose test suite links 38 binaries will hit this again on the next couple of
additions, and the fix for that is a CI config change (a larger runner, or
sharding the test job) rather than something to smuggle into a feature PR.

Verified: permission_ticket_test 14 passed and stable across three
consecutive runs, clippy -D warnings and rustfmt clean.
